refactor(pwdft): move init_vnl into vnl_pw_init_vnl.cpp
Step 5 of vnl_pw.cpp split:
- Move init_vnl() into new vnl_pw_init_vnl.cpp
- Replace new[]/delete[] with std::vector for jl/aux inside the tab fill loop
- Sync CMakeLists.txt and Makefile.Objects

Verified: make -j 30 in build_max_para_test passes.
