+++ source/Makefile.Objects
@@ -727,6 +727,7 @@ OBJS_SRCPW=h_ewald_pw.o\
     vnl_pw_deeq.o\
     vnl_pw_getvnl.o\
     vnl_pw_grad.o\
+    vnl_pw_init_vnl.o\
     vnl_pw_qrad.o\
     chgmixing.o\
     charge.o\
